refactor: Move utility function getIcon to Buttons
hence this will solve  import cycle problem
