+++ src/components/Icons/__tests__/utils.test.ts
@@ -1,0 +1,14 @@
+import { colors } from '@magnetis/astro-galaxy-tokens';
+
+import { getFill } from '../utils';
+
+describe('Icons/Utils', () => {
+  describe('getFill', () => {
+    it('when gradient is valid', () => {
+      expect(getFill({ gradient: 'sombrero', color: colors.uranus500, id: 'ID' })).toEqual('url(#ID)');
+    });
+    it('when gradient is not valid', () => {
+      expect(getFill({ gradient: undefined, color: colors.uranus500, id: 'ID' })).toEqual(colors.uranus500);
+    });
+  });
+});
