+++ src/components/Icons/utils.ts
@@ -1,7 +1,5 @@
 import { GradientID } from '@tokens/gradients';
 
-import * as icons from '@components/Icons';
-
 interface getFillParams {
   gradient?: GradientID;
   color: string;
@@ -13,14 +11,4 @@ export function getFill({ gradient, color, id }: getFillParams) {
     return `url(#${id})`;
   }
   return color;
-}
-
-export function getIcon(_iconName: string) {
-  const iconName = `${_iconName}Icon`;
-
-  if (Object.keys(icons).includes(iconName)) {
-    return (icons as { [key: string]: Function })[iconName];
-  }
-
-  return () => null;
 }
